Handled the currentDatabase and created an object to get the currentDatabase.
Name of teh currentDatabase is stored in a txt file which will bee read by the Currentdb Object to get the name of teh currentDatabase. Creating a new database will change the file content to the name of the new database. Now will have to implement a SQL command to change the selected database(currentdatabase)