Added a regex statement to detect and create databases and also change the default value of database in insert function.
Now this regex statement can detect the queries for creating database and then calls the createDatabase function to create database. Also changed the default value of database argument in the insert function to Currentdb.get() for the proper handling of current database. Also restructured the regex statement to remove white space mishandling.